feat: Moved to OfferContent isolated component
Introduce the OfferContent component to encapsulate the upsell offer content and functionality. This allows for better code organization and separation of concerns.

The changes include:
- Moved the product list, order summary, and call-to-action functionality from the UpsellPage component into the new OfferContent component.
- Removed the unused Countdown and OrderSummary components.
- Updated the UpsellPage component to use the new OfferContent component.

This change improves the overall code structure and makes the UpsellPage component more focused on its primary responsibilities, while the OfferContent component handles the specific upsell offer-related logic.
