+++ src/app/upsell/page.tsx
@@ -1,17 +1,14 @@
 "use client";
 
 import { useState, useEffect, useMemo } from "react";
-import { Countdown } from "../components/Countdown";
 import { LuckyCards } from "../components/LuckyCards";
 import { CroTools } from "../components/CroTools";
 import { UpsellData } from "@/types/UpsellData";
 import { Product } from "@/types/Product";
-import { ProductList } from "./components/ProductList";
-import { OrderSummary } from "./components/OrderSummary";
 import { OrderTotals } from "@/types/OrderTotals";
 import { PreHeader } from "./components/PreHeader";
-import Link from "next/link";
 import { useRouter, useSearchParams } from "next/navigation";
+import { OfferContent } from "./components/OfferContent";
 
 export default function UpsellPage() {
   const [data, setData] = useState<UpsellData | null>(null);
@@ -159,67 +156,3 @@ export default function UpsellPage() {
     </>
   );
 }
-
-export const OfferContent = ({
-  data,
-  selectedProducts,
-  totals,
-  toggleProductSelection,
-  handleAddToCart,
-}: {
-  data: UpsellData;
-  selectedProducts: number[];
-  totals: OrderTotals;
-  toggleProductSelection: (productId: number) => void;
-  handleAddToCart: () => void;
-}) => {
-  return (
-    <>
-      <h2 className="text-xl md:text-2xl font-bold text-gray-800 mb-4 mt-8 text-center">
-        {data.copy}
-      </h2>
-      <p className="text-center text-gray-600 mb-4">{data.subcopy}</p>
-      <Countdown />
-      <div className="space-y-8">
-        {/* Products Grid */}
-        <ProductList
-          products={data.products}
-          selectedProducts={selectedProducts}
-          toggleProductSelection={toggleProductSelection}
-        />
-
-        {/* Order Summary */}
-        {selectedProducts.length > 0 && <OrderSummary totals={totals} />}
-
-        {/* Call to Action */}
-        <div className="space-y-4 mt-8">
-          <button
-            onClick={() => {
-              handleAddToCart();
-            }}
-            className={`w-full py-4 mb-4 ${
-              selectedProducts.length > 0
-                ? "bg-blue-500 hover:bg-blue-600 min-h-[90px]"
-                : "bg-gray-400"
-            } text-white font-bold rounded-lg transition duration-200 text-lg shadow-md cursor-pointer`}
-            disabled={selectedProducts.length === 0}
-          >
-            {selectedProducts.length > 0
-              ? `${data.primaryCTA}`
-              : "Select at least one product to continue"}
-          </button>
-          <Link href="/thankyou">
-            <button className="w-full py-4 bg-white border border-gray-200 hover:bg-gray-300 text-gray-500 font-bold rounded-lg transition duration-200 text-lg shadow-md cursor-pointer">
-              {data.secondaryCTA}
-            </button>
-          </Link>
-        </div>
-
-        {/* Guarantee */}
-        <div className="text-center text-gray-500 text-sm">
-          {data.guarantee}
-        </div>
-      </div>
-    </>
-  );
-};
